// src/commom/types/DefaultMessages.ts
export enum DefaultMessages {
  QUERY_SUCCESS = 'Consulta realizada com sucesso.',
  CREATED = 'Criado com sucesso.',
  DATA_NOT_FOUND = 'Recurso não encontrado',
  DELETED = 'Deletado com sucesso.',
  UPDATED = 'Atualizado com sucesso.',
  EMAIL_OR_PASS_INVALID = 'E-mail e/ou senha incorretos.',
  TOKEN_NOT_FOUND = 'Token de autorização ausente ou inválido',
  NOT_AUTHORIZED = 'Você não possui permissão',
  INTERNAL_SERVER_ERROR = 'Ocorreu um erro desconhecido, contacte nossa equipe de suporte',
}